# Avoid growing copies of collections of known size