--- a/codegen/src/main/java/software/amazon/awssdk/codegen/poet/model/MemberCopierSpec.java
+++ b/codegen/src/main/java/software/amazon/awssdk/codegen/poet/model/MemberCopierSpec.java
@@ -269,18 +269,27 @@ class MemberCopierSpec implements ClassSpec {
             code.add("$T $N;", listType, outputVariableName)
                 .add("if ($1N == null || $1N instanceof $2T) {", inputVariableName, SdkAutoConstructList.class)
                 .add("$N = $T.getInstance();", outputVariableName, DefaultSdkAutoConstructList.class)
-                .add("} else {")
-                .add("$T $N = new $T<>();", listType, modifiableVariableName, ArrayList.class);
+                .add("} else {");
+
 
             String entryInputVariable = variableSource.getNew("entry");
-            code.add("$N.forEach($N -> {", inputVariableName, entryInputVariable);
 
-            String entryOutputVariable =
-                copyMethodBody(code, builderTransform, variableSource, enumTransform, entryInputVariable, listEntryModel);
+            // Short-circuit, if the member is simple then we can directly use the constructor
+            // that takes a collection which should have a better performance characteristics.
+            boolean isMemberSimple = listEntryModel.isSimple() && listEntryModel.getEnumType() == null;
+            if (isMemberSimple) {
+                code.add("$T $N = new $T<>($N);", listType, modifiableVariableName, ArrayList.class, inputVariableName);
+            } else {
+                code.add("$T $N = new $T<>($N.size());", listType, modifiableVariableName, ArrayList.class, inputVariableName);
+                code.add("$N.forEach($N -> {", inputVariableName, entryInputVariable);
 
-            code.add("$N.add($N);", modifiableVariableName, entryOutputVariable)
-                .add("});")
-                .add("$N = $T.unmodifiableList($N);", outputVariableName, Collections.class, modifiableVariableName)
+                String entryOutputVariable =
+                    copyMethodBody(code, builderTransform, variableSource, enumTransform, entryInputVariable, listEntryModel);
+
+                code.add("$N.add($N);", modifiableVariableName, entryOutputVariable)
+                    .add("});");
+            }
+            code.add("$N = $T.unmodifiableList($N);", outputVariableName, Collections.class, modifiableVariableName)
                 .add("}");
 
             return outputVariableName;
@@ -299,7 +308,8 @@ class MemberCopierSpec implements ClassSpec {
                 .add("if ($1N == null || $1N instanceof $2T) {", inputVariableName, SdkAutoConstructMap.class)
                 .add("$N = $T.getInstance();", outputVariableName, DefaultSdkAutoConstructMap.class)
                 .add("} else {")
-                .add("$T $N = new $T<>();", outputMapType, modifiableVariableName, LinkedHashMap.class);
+                .add("$T $N = new $T<>($N.size());",
+                     outputMapType, modifiableVariableName, LinkedHashMap.class, inputVariableName);
 
             String keyInputVariable = variableSource.getNew("key");
             String valueInputVariable = variableSource.getNew("value");