--- a/codegen/src/test/resources/software/amazon/awssdk/codegen/poet/model/listofblobstypecopier.java
+++ b/codegen/src/test/resources/software/amazon/awssdk/codegen/poet/model/listofblobstypecopier.java
@@ -18,10 +18,7 @@ final class ListOfBlobsTypeCopier {
         if (listOfBlobsTypeParam == null || listOfBlobsTypeParam instanceof SdkAutoConstructList) {
             list = DefaultSdkAutoConstructList.getInstance();
         } else {
-            List<SdkBytes> modifiableList = new ArrayList<>();
-            listOfBlobsTypeParam.forEach(entry -> {
-                modifiableList.add(entry);
-            });
+            List<SdkBytes> modifiableList = new ArrayList<>(listOfBlobsTypeParam);
             list = Collections.unmodifiableList(modifiableList);
         }
         return list;