--- a/codegen/src/test/resources/software/amazon/awssdk/codegen/poet/model/listoflistoflistofstringscopier.java
+++ b/codegen/src/test/resources/software/amazon/awssdk/codegen/poet/model/listoflistoflistofstringscopier.java
@@ -13,27 +13,24 @@ import software.amazon.awssdk.core.util.SdkAutoConstructList;
 @Generated("software.amazon.awssdk:codegen")
 final class ListOfListOfListOfStringsCopier {
     static List<List<List<String>>> copy(
-        Collection<? extends Collection<? extends Collection<String>>> listOfListOfListOfStringsParam) {
+            Collection<? extends Collection<? extends Collection<String>>> listOfListOfListOfStringsParam) {
         List<List<List<String>>> list;
         if (listOfListOfListOfStringsParam == null || listOfListOfListOfStringsParam instanceof SdkAutoConstructList) {
             list = DefaultSdkAutoConstructList.getInstance();
         } else {
-            List<List<List<String>>> modifiableList = new ArrayList<>();
+            List<List<List<String>>> modifiableList = new ArrayList<>(listOfListOfListOfStringsParam.size());
             listOfListOfListOfStringsParam.forEach(entry -> {
                 List<List<String>> list1;
                 if (entry == null || entry instanceof SdkAutoConstructList) {
                     list1 = DefaultSdkAutoConstructList.getInstance();
                 } else {
-                    List<List<String>> modifiableList1 = new ArrayList<>();
+                    List<List<String>> modifiableList1 = new ArrayList<>(entry.size());
                     entry.forEach(entry1 -> {
                         List<String> list2;
                         if (entry1 == null || entry1 instanceof SdkAutoConstructList) {
                             list2 = DefaultSdkAutoConstructList.getInstance();
                         } else {
-                            List<String> modifiableList2 = new ArrayList<>();
-                            entry1.forEach(entry2 -> {
-                                modifiableList2.add(entry2);
-                            });
+                            List<String> modifiableList2 = new ArrayList<>(entry1);
                             list2 = Collections.unmodifiableList(modifiableList2);
                         }
                         modifiableList1.add(list2);