--- a/codegen/src/test/resources/software/amazon/awssdk/codegen/poet/model/mapofenumtoenumcopier.java
+++ b/codegen/src/test/resources/software/amazon/awssdk/codegen/poet/model/mapofenumtoenumcopier.java
@@ -16,7 +16,7 @@ final class MapOfEnumToEnumCopier {
         if (mapOfEnumToEnumParam == null || mapOfEnumToEnumParam instanceof SdkAutoConstructMap) {
             map = DefaultSdkAutoConstructMap.getInstance();
         } else {
-            Map<String, String> modifiableMap = new LinkedHashMap<>();
+            Map<String, String> modifiableMap = new LinkedHashMap<>(mapOfEnumToEnumParam.size());
             mapOfEnumToEnumParam.forEach((key, value) -> {
                 modifiableMap.put(key, value);
             });
@@ -30,7 +30,7 @@ final class MapOfEnumToEnumCopier {
         if (mapOfEnumToEnumParam == null || mapOfEnumToEnumParam instanceof SdkAutoConstructMap) {
             map = DefaultSdkAutoConstructMap.getInstance();
         } else {
-            Map<String, String> modifiableMap = new LinkedHashMap<>();
+            Map<String, String> modifiableMap = new LinkedHashMap<>(mapOfEnumToEnumParam.size());
             mapOfEnumToEnumParam.forEach((key, value) -> {
                 String result = key.toString();
                 String result1 = value.toString();
@@ -46,7 +46,7 @@ final class MapOfEnumToEnumCopier {
         if (mapOfEnumToEnumParam == null || mapOfEnumToEnumParam instanceof SdkAutoConstructMap) {
             map = DefaultSdkAutoConstructMap.getInstance();
         } else {
-            Map<EnumType, EnumType> modifiableMap = new LinkedHashMap<>();
+            Map<EnumType, EnumType> modifiableMap = new LinkedHashMap<>(mapOfEnumToEnumParam.size());
             mapOfEnumToEnumParam.forEach((key, value) -> {
                 EnumType result = EnumType.fromValue(key);
                 EnumType result1 = EnumType.fromValue(value);